aichatbot-langchain/requirements.txt
# YOLOv7 Acne Detection API
# Python dependencies for running the inference server

fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19
pillow==11.0.0
torch==2.5.1
torchvision==0.20.1
opencv-python==4.10.0.84
numpy==2.1.3
pydantic==2.10.3
python-dotenv==1.0.1
ultralytics>=8.0.0
pyyaml>=6.0
